backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
PyJWT>=2.7.0
pydantic>=2.0.0
docker>=7.0.0
